+++ apps/web/app/(ee)/api/stripe/integration/webhook/invoice-paid.ts
@@ -5,7 +5,7 @@ import { includeTags } from "@/lib/api/links/include-tags";
 import { syncPartnerLinksStats } from "@/lib/api/partners/sync-partner-links-stats";
 import { executeWorkflows } from "@/lib/api/workflows/execute-workflows";
 import { createPartnerCommission } from "@/lib/partners/create-partner-commission";
-import { sendPartnerPostback } from "@/lib/postback/api/send-partner-postback";
+import { sendPartnerPostback } from "@/lib/postback/send-partner-postback";
 import { getLeadEvent, recordSale } from "@/lib/tinybird";
 import { StripeMode } from "@/lib/types";
 import { redis } from "@/lib/upstash";
