+++ apps/web/lib/partners/create-partner-commission.ts
@@ -18,7 +18,7 @@ import { getProgramEnrollmentOrThrow } from "../api/programs/get-program-enrollm
 import { calculateSaleEarnings } from "../api/sales/calculate-sale-earnings";
 import { executeWorkflows } from "../api/workflows/execute-workflows";
 import { Session } from "../auth";
-import { sendPartnerPostback } from "../postback/api/send-partner-postback";
+import { sendPartnerPostback } from "../postback/send-partner-postback";
 import { RewardContext, RewardProps } from "../types";
 import { sendWorkspaceWebhook } from "../webhook/publish";
 import { CommissionWebhookSchema } from "../zod/schemas/commissions";
