+++ apps/web/lib/postback/get-postback-events.ts
@@ -1,6 +1,6 @@
 import { postbackEventOutputSchemaTB } from "@/lib/postback/schemas";
 import * as z from "zod/v4";
-import { tb } from "../../tinybird/client";
+import { tb } from "../tinybird/client";
 
 export const getPostbackEvents = tb.buildPipe({
   pipe: "get_postback_events",
