+++ apps/web/lib/postback/record-postback-event.ts
@@ -1,5 +1,5 @@
 import { postbackEventInputSchemaTB } from "@/lib/postback/schemas";
-import { tb } from "../../tinybird/client";
+import { tb } from "../tinybird/client";
 
 export const recordPostbackEvent = tb.buildIngestEndpoint({
   datasource: "dub_postback_events",
